feat: 닉네임 중복 체크 추가
